feat(scalper): 支持小数张开仓，按lotSz精度取整而非强制1张
- OKX合约支持0.01张精度，160 USDT预算下BTC/ETH/SOL均可开仓
  之前取整为整数张导致小资金下BTC/ETH完全开不起
- calc_contracts 新增 lot_sz 参数，按 lotSz 网格向下取整（默认0.01）
- broker 新增 lot_size() 从市场信息读取最小下单量
- 开/平仓 contracts 类型 int → float，DB 列 INTEGER → REAL
- 160 USDT 实际效果：BTC 0.12张 / ETH 0.41张 / SOL 1.04张，均等额~78 USDT名义
- 33个单元测试全过

+++ tradingagents/scalper/state_store.py
@@ -32,7 +32,7 @@ class ScalperStore:
                 CREATE TABLE IF NOT EXISTS scalper_positions (
                     symbol TEXT PRIMARY KEY,
                     side TEXT NOT NULL,
-                    contracts INTEGER NOT NULL,
+                    contracts REAL NOT NULL,
                     entry_price REAL NOT NULL,
                     margin REAL NOT NULL,
                     entry_time REAL NOT NULL,
@@ -56,7 +56,7 @@ class ScalperStore:
                     side TEXT NOT NULL,
                     action TEXT NOT NULL,
                     price REAL NOT NULL,
-                    contracts INTEGER NOT NULL,
+                    contracts REAL NOT NULL,
                     margin REAL,
                     pnl_usdt REAL,
                     pnl_pct REAL,
